# docs/specs/purchase-sales/openapi.yaml
openapi: 3.0.3
info:
  title: Fasse 仕入・売上管理API
  version: "1.0.0"
  description: |
    仕入・売上管理のAPI仕様（APIGateway - Lambda - DynamoDBで提供）。詳細はdesign.mdを参照。
    全エンドポイントでJWT認証（Authorization: Bearer <JWT>）が必須。詳細はdocs/specs/authenticationを参照。

    リクエストの扱い（詳細はdesign.md「入力検証・エラー応答」）:
    - リクエストボディ・パラメータは本仕様のスキーマで検証し、満たさない場合は400を返す
    - スキーマに定義されていない項目、およびサーバーが管理する項目（id、purchase_no/sales_no、created_at、updated_at）はリクエストに含まれていても無視する
    - PUTで任意項目を省略した場合は既存の値を維持する
    - created_at/updated_atはUTCのISO 8601文字列。業務日付・日時（purchase_date、business_date、sales_datetime等）はJST固定
servers:
  - url: https://{apiId}.execute-api.{region}.amazonaws.com/stg
    variables:
      apiId:
        default: apiId
        description: API GatewayのREST API ID（デプロイ後に確認する）
      region:
        default: ap-northeast-1
security:
  - bearerAuth: []
paths:
...
  /tax-rates/{taxCategory}/{validFrom}:
    parameters:
      - $ref: "#/components/parameters/TaxCategoryPath"
      - $ref: "#/components/parameters/ValidFrom"
    get:
      summary: 消費税率取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRate"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: 消費税率更新（tax_category・valid_fromはパスの値が優先され、リクエストボディの値は無視される。税率改定時はvalid_toの設定＋新期間のPOSTで対応するのが基本運用）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxRateUpdateInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRate"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: 消費税率削除（物理削除。仕入/売上明細はtax_rateを数値でスナップショット保持しFK参照しないため、削除しても既存伝票に影響しない）
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
...
    TaxCategoryPath:
      name: taxCategory
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/TaxCategory"
    ValidFrom:
      name: validFrom
      in: path
      required: true
      schema:
        type: string
        format: date
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT発行API（POST /auth/token、POST /auth/token/cognito）で取得したKMS署名のJWT。詳細はdocs/specs/authenticationを参照
  responses:
    BadRequest:
      description: リクエストがスキーマを満たさない（messageに不備のある項目名を含む）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: JWTが無い、形式・署名が不正、または有効期限切れ
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: 予期しないエラー（messageは固定文言。詳細はログにのみ出力する）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 利用者向けのメッセージ
        requestId:
          type: string
          description: 問い合わせ・ログ調査用のリクエストID
...
    TaxCategory:
      type: string
      enum: [STANDARD, REDUCED, EXEMPT]
      description: 標準税率/軽減税率/非課税。消費税法で定まる固定区分のため独立マスタにはせず、ENUM値として持つ
    TaxRateInput:
      type: object
      required: [tax_category, description, rate, valid_from]
      properties:
        tax_category:
          $ref: "#/components/schemas/TaxCategory"
        description:
          type: string
          description: "税区分の概要（例: 標準税率, 軽減税率(食品等), 非課税）"
        rate:
          type: number
          description: "税率（例: 0.10, 0.08, 0）"
        valid_from:
          type: string
          format: date
        valid_to:
          type: string
          format: date
          description: 未設定の場合は現在も有効
    TaxRate:
      allOf:
        - $ref: "#/components/schemas/TaxRateInput"
        - type: object
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    TaxRateUpdateInput:
      type: object
      required: [description, rate]
      description: tax_category・valid_fromはパスパラメータで指定するため含めない
      properties:
        description:
          type: string
        rate:
          type: number
        valid_to:
          type: string
          format: date